Extend track file format to allow to re-use setup from spec files
The spec files in the track file can now be grouped into `full´,
`fixtures` and `queries` to declare which parts of a spec should be
executed.
This is a bit of a trade-off to enable users to create a shareable setup
across multiple spec files without sacrificing the ability to run spec
files isolated.

Closes https://github.com/mfussenegger/cr8/issues/94