Insert-fake-data: Fix data generation to avoid duplicates
This replaces the ProcessPoolExecutor with a ThreadPoolExecutor.

Using multiprocessing leads to duplicate random data. This was the case
because each subprocess has a new random instance but inherits the seed
from the parent.

This was always a problem, but due to a change in Faker
(https://github.com/joke2k/faker/commit/7229f423cb7c41e1bd1332ba96fa73bde348a1a6)
it's now more of an issue.

This is a quick fix - until a solution that works with multiprocessing
is found.